+++ CHANGELOG.md
@@ -4,7 +4,11 @@ FEATURES:
 
 * **Run Pack from Folder** - Nomad Pack can run and render packs stored in the current folder. For example, if the current folder
   contains a pack named `simple-service`, you can run it using `nomad-pack run ./simple-service`.
-  
+
+* **Set Nomad Pack variables using environment variables** - Pack reads the
+  environment for variables prefixed with `NOMAD_PACK_VAR_` and supplies them to
+  the running pack.
+
 BUG FIXES:
 * template: Handle HEREDOC delimiter immediately before EOF [[GH-191](https://github.com/hashicorp/nomad-pack/pull/191)]
 * cli: display API client errors in CLI output [[GH-183](https://github.com/hashicorp/nomad-pack/pull/183)]